@@ -8,30 +8,30 @@
     "type": "git",
     "url": "https://github.com/cygni/snakebot-client-js.git"
   },
-  "module": "./src/index.js",
-  "browser": "./src/index.js",
-  "bin": {
-    "snake": "./bin/snake"
-  },
+  "private": true,
+  "type": "module",
+  "main": "index.js",
+  "browser": "src/index.js",
   "scripts": {
     "lint": "eslint .",
     "format": "prettier --write **/*.{js,json,md}",
-    "test": "mocha --require esm **/*.test.js",
-    "start": "node ./bin/snake"
+    "test": "mocha --experimental-modules",
+    "start": "node --experimental-modules --unhandled-rejections=strict snake-cli.js"
   },
   "dependencies": {
-    "commander": "^2.19.0",
-    "esm": "^3.2.16",
-    "ws": "^6.2.0"
+    "commander": "^4.1.1",
+    "ws": "^7.2.1"
   },
   "devDependencies": {
-    "@types/mocha": "^5.2.6",
-    "@types/node": "^10.14.1",
-    "eslint": "^5.13.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-prettier": "^3.0.1",
-    "mocha": "^6.0.0",
-    "prettier": "^1.16.4"
+    "@types/mocha": "^7.0.1",
+    "@types/node": "^12.12.6",
+    "@types/ws": "^7.2.1",
+    "babel-eslint": "^10.0.3",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-prettier": "^3.1.2",
+    "mocha": "7.0.0-esm1",
+    "prettier": "^1.19.1"
   }
 }
