@@ -5,54 +5,48 @@ const allDirections = Object.values(Direction);
 
 /**
  * @template T
- * @param {ArrayLike<T>} items
+ * @param {T[]} items
  * @returns {T}
  */
 function randomItem(items) {
   return items[Math.floor(Math.random() * items.length)];
 }
 
-export default class Slither {
-  constructor() {
-    // The snake must have a name
-    this.name = 'Slither';
+export const SNAKE_NAME = 'Slither';
+
+/**
+ * @param {import('../utils').GameMap} gameMap
+ * @returns {Direction}
+ */
+export function getNextMove(gameMap) {
+  const currentCoordinate = gameMap.playerSnake.headCoordinate;
+
+  const safeDirections = allDirections.filter(direction => {
+    const nextCoordinate = currentCoordinate.translatedByDirection(direction);
+    const nextTile = gameMap.getTile(nextCoordinate);
+
+    switch (nextTile.type) {
+      case TileType.Empty:
+      case TileType.Food:
+        return true;
+      default:
+        return false;
+    }
+  });
+
+  // Bad luck!
+  if (safeDirections.length === 0) {
+    return Direction.Down;
   }
 
-  // This method is required
-  /**
-   * @param {import('../utils').GameMap} gameMap
-   * @returns {Direction}
-   */
-  getNextMove(gameMap) {
-    const currentCoordinate = gameMap.playerSnake.headCoordinate;
+  return randomItem(safeDirections);
+}
 
-    const safeDirections = allDirections.filter(direction => {
-      const nextCoordinate = currentCoordinate.translatedByDirection(direction);
-      const nextTile = gameMap.getTile(nextCoordinate);
-
-      switch (nextTile.type) {
-        case TileType.Empty:
-        case TileType.Food:
-          return true;
-        default:
-          return false;
-      }
-    });
-
-    // Bad luck!
-    if (safeDirections.length === 0) {
-      return Direction.Down;
-    }
-
-    return randomItem(safeDirections);
-  }
-
-  // This method is optional
-  onMessage(message) {
-    switch (message.type) {
-      case MessageType.GameStarting:
-        // Reset snake state here
-        break;
-    }
+// This handler is optional
+export function onMessage(message) {
+  switch (message.type) {
+    case MessageType.GameStarting:
+      // Reset snake state here
+      break;
   }
 }
